Syncing Pymbar4 with master

+++ pymbar/testsystems/exponential_distributions.py
@@ -65,13 +65,13 @@ class ExponentialTestCase(object):
         return np.log(self.rates)
 
     def analytical_means(self):
-        return self.rates ** -1.0
+        return self.rates**-1.0
 
     def analytical_variances(self):
-        return self.rates ** -2.0
+        return self.rates**-2.0
 
     def analytical_standard_deviations(self):
-        return np.sqrt(self.rates ** -2.0)
+        return np.sqrt(self.rates**-2.0)
 
     def analytical_observable(self, observable="position"):
 
@@ -170,7 +170,7 @@ class ExponentialTestCase(object):
         x_kn = np.zeros([self.n_states, N_max], np.float64)
         u_kln = np.zeros([self.n_states, self.n_states, N_max], np.float64)
         x_n = np.zeros([N_tot], np.float64)
-        s_n = np.zeros([N_tot], np.int)
+        s_n = np.zeros([N_tot], int)
         u_kn = np.zeros([self.n_states, N_tot], np.float64)
         index = 0
         for k, N in enumerate(N_k):
